cleanup: move stuff into serperate components
