+++ src/lib/components/ui/chat/chat-attachment.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+	import { Paperclip } from 'lucide-svelte';
+	import { Button } from '$lib/components/ui/button';
+</script>
+
+<Button class="rounded-sm" size="icon" disabled>
+	<Paperclip />
+</Button>
